Package commands and utils and update packages in other files and tests
