+++ src/main/java/ch/heigvd/app/utils/parsers/JsonConverter.java
@@ -1,0 +1,22 @@
+package ch.heigvd.app.utils.parsers;
+
+import com.google.gson.Gson;
+
+/**
+ * JsonConverter class
+ */
+public class JsonConverter {
+    /**
+     * Convert a json string into a java object
+     * @param input the json string
+     * @return the java object
+     * @see <a href="https://github.com/google/gson">gson</a>
+     */
+    public static JavaConfig convert(String input){
+        Gson gson = new Gson();
+        JavaConfig javaConf = gson.fromJson(input, JavaConfig.class);
+
+        return javaConf;
+    }
+}
+
