+++ src/test/java/ch/heigvd/app/utils/JsonConverterTest.java
@@ -1,5 +1,7 @@
 package ch.heigvd.app.utils;
 
+import ch.heigvd.app.utils.parsers.JavaConfig;
+import ch.heigvd.app.utils.parsers.JsonConverter;
 import org.junit.Test;
 
 import static org.junit.Assert.assertEquals;
@@ -19,9 +21,9 @@ public class JsonConverterTest {
                 "}";
         JavaConfig output = JsonConverter.convert(input);
 
-        assertEquals(result.title , output.title);
-        assertEquals(result.lang, output.lang);
-        assertEquals(result.charset, output.charset);
+        assertEquals(result.getTitle() , output.getTitle());
+        assertEquals(result.getLang(), output.getLang());
+        assertEquals(result.getCharset(), output.getCharset());
     }
 
 }
